refactor: 네이밍 및 패키지 구조 변경
domain 과 web 관련 클래스를 구분,
admin/product -> admin/products
