+++ src/main/java/cart/domain/admin/persistence/entity/ProductEntityMappingException.java
@@ -1,0 +1,12 @@
+package cart.domain.admin.persistence.entity;
+
+public class ProductEntityMappingException extends RuntimeException {
+
+    ProductEntityMappingException() {
+
+    }
+
+    ProductEntityMappingException(String message) {
+        super(message);
+    }
+}
