+++ src/main/java/cart/domain/admin/service/DbNotAffectedException.java
@@ -1,0 +1,12 @@
+package cart.domain.admin.service;
+
+public class DbNotAffectedException extends RuntimeException {
+
+    DbNotAffectedException() {
+
+    }
+
+    DbNotAffectedException(String message) {
+        super(message);
+    }
+}
