--- src/main/java/cart/persistence/dao/ProductDao.java
@@ -1,18 +1,0 @@
-package cart.persistence.dao;
-
-import java.util.List;
-
-import cart.persistence.entity.ProductEntity;
-
-public interface ProductDao {
-
-    Long save(ProductEntity productEntity);
-
-    ProductEntity findByName(String name);
-
-    List<ProductEntity> findAll();
-
-    int update(ProductEntity productEntity);
-
-    int deleteById(long id);
-}
