--- src/main/java/cart/persistence/entity/ProductEntityMappingException.java
@@ -1,12 +1,0 @@
-package cart.persistence.entity;
-
-public class ProductEntityMappingException extends RuntimeException {
-
-    ProductEntityMappingException() {
-
-    }
-
-    ProductEntityMappingException(String message) {
-        super(message);
-    }
-}
