--- src/main/java/cart/service/DbNotAffectedException.java
@@ -1,12 +1,0 @@
-package cart.service;
-
-public class DbNotAffectedException extends RuntimeException {
-
-    DbNotAffectedException() {
-
-    }
-
-    DbNotAffectedException(String message) {
-        super(message);
-    }
-}
