+++ src/test/java/cart/AdminApiEndToEndTest.java
@@ -17,10 +17,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.jdbc.Sql;
 
-import cart.dto.ProductPostRequest;
-import cart.dto.ProductPutRequest;
-import cart.persistence.dao.ProductDao;
-import cart.persistence.entity.ProductEntity;
+import cart.domain.admin.persistence.dao.ProductDao;
+import cart.domain.admin.persistence.entity.ProductEntity;
+import cart.web.admin.dto.PostProductRequest;
+import cart.web.admin.dto.PutProductRequest;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
@@ -44,12 +44,12 @@ class AdminApiEndToEndTest {
     }
 
     private ExtractableResponse<Response> saveProduct(final String name, final int price, final String imageUrl) {
-        final ProductPostRequest request = new ProductPostRequest(name, price, imageUrl);
+        final PostProductRequest request = new PostProductRequest(name, price, imageUrl);
 
         return given()
             .body(request)
             .when()
-            .post("/admin/product")
+            .post("/admin/products")
             .then()
             .extract();
     }
@@ -70,7 +70,7 @@ class AdminApiEndToEndTest {
 
             SoftAssertions.assertSoftly(softAssertions -> {
                 softAssertions.assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-                softAssertions.assertThat(response.header("Location")).contains("/admin/product/");
+                softAssertions.assertThat(response.header("Location")).contains("/admin/products/");
                 softAssertions.assertThat(savedEntity.getName()).isEqualTo("modi");
                 softAssertions.assertThat(savedEntity.getPrice()).isEqualTo(10000);
                 softAssertions.assertThat(savedEntity.getImageUrl()).isEqualTo("https://woowacourse.github.io/");
@@ -95,17 +95,17 @@ class AdminApiEndToEndTest {
             final String[] locations = response.header("Location").split("/");
             final String id = locations[locations.length - 1];
 
-            final ProductPostRequest productPostRequest = new ProductPostRequest("modi", 15000, "https://changed.com/");
+            final PostProductRequest postProductRequest = new PostProductRequest("modi", 15000, "https://changed.com/");
             given()
-                .body(productPostRequest)
-                .when().put("/admin/product/" + id)
+                .body(postProductRequest)
+                .when().put("/admin/products/" + id)
                 .then()
                 .statusCode(HttpStatus.OK.value());
 
             final ProductEntity changedEntity = productDao.findByName("modi");
             SoftAssertions.assertSoftly(softAssertions -> {
                 softAssertions.assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-                softAssertions.assertThat(response.header("Location")).contains("/admin/product/" + changedEntity.getId());
+                softAssertions.assertThat(response.header("Location")).contains("/admin/products/" + changedEntity.getId());
                 softAssertions.assertThat(changedEntity.getName()).isEqualTo("modi");
                 softAssertions.assertThat(changedEntity.getPrice()).isEqualTo(15000);
                 softAssertions.assertThat(changedEntity.getImageUrl()).isEqualTo("https://changed.com/");
@@ -119,7 +119,7 @@ class AdminApiEndToEndTest {
             final String id = locations[locations.length - 1];
 
             given()
-                .when().delete("/admin/product/" + id)
+                .when().delete("/admin/products/" + id)
                 .then()
                 .statusCode(HttpStatus.OK.value());
 
@@ -212,12 +212,12 @@ class AdminApiEndToEndTest {
             productDao.save(request);
 
             final Long wrongId = 0L;
-            final ProductPutRequest putRequest = new ProductPutRequest("modi", 7770, "https://woowacourse.github.io/");
+            final PutProductRequest putRequest = new PutProductRequest("modi", 7770, "https://woowacourse.github.io/");
 
             final ExtractableResponse<Response> response = given()
                 .body(putRequest)
                 .when()
-                .put("/admin/product/" + wrongId)
+                .put("/admin/products/" + wrongId)
                 .then()
                 .statusCode(HttpStatus.NOT_FOUND.value())
                 .extract();
@@ -231,7 +231,7 @@ class AdminApiEndToEndTest {
 
             final ExtractableResponse<Response> response = given()
                 .when()
-                .delete("/admin/product/" + wrongId)
+                .delete("/admin/products/" + wrongId)
                 .then()
                 .statusCode(HttpStatus.NOT_FOUND.value())
                 .extract();
